refactor: rename State to GameManager
