+++ poketechnika/Game.h
@@ -1,15 +1,14 @@
 #pragma once
 #include <SFML/Graphics.hpp>
-#include "State.h"
+#include "GameManager.h"
 class Game
 {
 public:
 	Game();
 	void InitGameLoop(int width, int height);
-	State state_;
+	GameManager gm;
 private:
 	void ProcessInput(sf::RenderWindow& window);
 	void GameUpdate();
 	void Render(sf::RenderWindow& window);
-
 };
